% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMbehavr.R
\name{HMMbehavr}
\alias{HMMbehavr}
\title{HMMbehavr}
\usage{
HMMbehavr(behavtbl, it = 100, ldcyc = NULL)
}
\arguments{
\item{behavtbl}{A data frame containing the behavioral data, conforming to a \code{behavr} table.}

\item{it}{Integer (>= 100) specifying the number of iterations for the HMM fitting process. Defaults to 100.}

\item{ldcyc}{A single numeric value specifying the duration of the light phase in hours. If NULL (default), a 12-hour light/12-hour dark cycle is assumed.}
}
\value{
A list containing data frames:
\item{TimeSpentInEachState}{Time spent in each inferred state.}
\item{VITERBIDecodedProfile}{The most likely sequence of hidden states (Viterbi path).}
}
\description{
This function runs a Hidden Markov Model (HMM) on behavioral data.
}
\examples{
res1 <- HMMbehavr(behavtbl = dt_test, it = 100, ldcyc = 10) # 10 hours light, 100 iterations
}
