% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMbehavrFast.R
\name{HMMbehavrFast}
\alias{HMMbehavrFast}
\title{Run Hidden Markov Model (HMM) on behavioural data in parallel}
\usage{
HMMbehavrFast(behavtbl, it = 100, n_cores = 4, ldcyc = NULL)
}
\arguments{
\item{behavtbl}{A data frame containing the behavioral data. This data frame must conform to the structure of a \code{behavr} table.}

\item{it}{Integer (>= 100) specifying the number of iterations for the HMM fitting process. Defaults to 100.}

\item{n_cores}{An integer specifying the number of CPU cores to use for parallel processing. Defaults to 4.}

\item{ldcyc}{A single numeric value specifying the duration of the light phase in hours. If NULL (default), a 12-hour light/12-hour dark cycle is assumed.}
}
\value{
A list containing two data frames:
\itemize{
\item{\code{TimeSpentInEachState}: Time spent in each inferred state for each individual.}
\item{\code{VITERBIDecodedProfile}: HMM-inferred state profiles for each individual over time.}
}
}
\description{
This function applies a Hidden Markov Model (HMM) to behavioral data for multiple individuals in parallel using the \code{foreach} package. It serves as a wrapper for the \code{HMMbehavr} function.
}
\examples{
\dontrun{
# Assuming 'dt_test' is a valid behavr table
results <- HMMbehavrFast(behavtbl = dt_test, iterations = 100, n_states = 4, n_cores = 10, ldcyc = 12, max_inner_it = 500)
}
}
