% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMplot.R
\name{HMMplot}
\alias{HMMplot}
\title{Plot Inferred Sleep States from HMM}
\usage{
HMMplot(hmm_inference_list, color_palette = "default", user_colors = NULL)
}
\arguments{
\item{hmm_inference_list}{A list, typically the output of the \code{HMMbehavr} function. It is expected to contain a data frame at the second position (\verb{[[2]]}) with columns like \code{timestamp}, \code{ID}, \code{state_name}, and \code{day}.}

\item{color_palette}{Name of the desired color palette. Available options are \code{"default"}, \code{"AG"}, and \code{"user"}. Defaults to \code{"default"}.}

\item{user_colors}{A named character vector specifying the colors for each state (e.g., \code{c("State0" = "red", "State1" = "blue")}).
This parameter is only used when \code{color_palette = "user"}. If \code{"user"} is selected and this parameter is not provided or is incorrectly formatted, an error will be thrown.}
}
\value{
A \code{ggplot2} object representing the plot of inferred sleep states.
}
\description{
This function visualizes the sleep state flags inferred by a Hidden Markov Model (HMM) applied to behavioral data.
}
\examples{
# Assuming 'res1' is the output of HMMbehavr
library(ggplot2)

# Using default palette
plot_default <- HMMplot(hmm_inference_list = res1)
print(plot_default)

# Using AG palette
plot_AG <- HMMplot(hmm_inference_list = res1, color_palette = "AG")
print(plot_AG)

# Using a custom user-defined palette
user_palette <- c("State0" = "lightgreen", "State1" = "darkgreen", "State2" = "lightblue", "State3" = "darkblue")
plot_user <- HMMplot(hmm_inference_list = res1, color_palette = "user", user_colors = user_palette)
print(plot_user)
}
