% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmmSinglePlot.R
\name{HMMSinglePlot}
\alias{HMMSinglePlot}
\title{HMMplot}
\usage{
HMMSinglePlot(HMMinferList, col_palette = "default")
}
\arguments{
\item{HMMinferList}{A list. Output of \code{HMMbehavr}.}

\item{col_palette}{Name of desired color palette. Currently 2 options available - \code{default}, and \code{AG}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function is used to save the sleep states inferred through a Hidden Markov Model (HMM) on behavioral data.
The function saves the plots in the current working directory.
}
\examples{
HMMSinglePlot(HMMinferList = res1)
}
