% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepDAMAnnotation.R
\name{sleepDAMAnnotation}
\alias{sleepDAMAnnotation}
\title{Custom Sleep Annotation Function}
\usage{
sleepDAMAnnotation(
  data,
  min_time_immobile = c(behavr::mins(5), behavr::mins(1440))
)
}
\arguments{
\item{data}{A data.table containing activity data for a single animal.
Must include 'activity' and 't' columns.}

\item{min_time_immobile}{A vector of two durations defining the min and max
length of an immobility bout to be considered sleep.}
}
\value{
The input data.table with two new columns: 'moving' (boolean) and
'asleep' (boolean).
}
\description{
Annotates sleep bouts based on immobility duration.
This function is designed to be used with \code{damr::load_dam}.
This function is modified from \code{sleepr::sleep_dam_annotation} function.
}
\keyword{internal}
